misc/HtmlTestSuite/results/test-0981.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
  <title>Deferred loading=lazy images load relative to the document's base URL
         at parse-time</title>
  <link href="mailto:dom@chromium.org" rel="author" title="Dom Farolino"/>
  <link href="mailto:rajendrant@chromium.org" rel="author" title="Raj T"/>
  <script src="/resources/testharness.js"/>
  <script src="/resources/testharnessreport.js"/>
  <script src="../resources/common.js"/>
<script>
  const below_viewport_img = new ElementLoadPromise("below-viewport");

  let has_window_loaded = false;

  async_test(t =&gt; {
    // Change the document's base URL to a bogus one, and scroll the
    // below-viewport img into view. When it loads, it should load relative
    // to the old base URL computed at parse-time.
    window.addEventListener("load", t.step_func(() =&gt; {
      window.history.pushState(2, document.title,
                               '/invalid-url-where-no-subresources-exist/')
      has_window_loaded = true;
      below_viewport_img.element().scrollIntoView();
    }));

    below_viewport_img.promise.then(t.step_func_done(() =&gt; {
      assert_true(has_window_loaded,
                  "Below-viewport loading=lazy images do not block the " +
                  "window load event");
    }));

    below_viewport_img.promise.catch(
      t.unreached_func("The image request should not load relative to the " +
                       "current (incorrect) base URL.")
    );
  }, "When a loading=lazy image is loaded, it loads relative to the " +
     "document's base URL computed at parse-time.");
</script></head>



<body>
  <div style="height:1000vh;"/>
  <script>
    // Change the document's base URL so that the img request parses relative
    // to it when it sets up the request at parse-time.
    window.history.pushState(1, document.title, 'resources/')
  </script>
  <img id="below-viewport" loading="lazy" onerror="below_viewport_img.reject()" onload="below_viewport_img.resolve()" src="image.png?base-url-2"/>
</body>
</html>
